Add new packages and improve font installation script
- Added `libgnome-keyring` and `ttf-jetbrains-mono-nerd` to the Arch packages installation script.
- Enhanced the font installation script to download the latest versions of JetBrains Mono Nerd Font and Cascadia Code, with a check for non-Arch systems.
- Added two new wallpapers to the Pictures directory.

+++ run_once_60-fonts.sh
@@ -4,18 +4,26 @@ echo "______________________"
 echo "Installing fonts"
 echo "======================"
 
-mkdir -p ~/.local/share/fonts
+if [[ ! -f /etc/arch-release ]]; then
+  echo "System is not Arch, installing fonts"
+  mkdir -p ~/.local/share/fonts
 
-# JetBrains Mono
-wget -O /tmp/JetBrainsMono.zip https://github.com/JetBrains/JetBrainsMono/releases/download/v2.242/JetBrainsMono-2.242.zip
-unzip /tmp/JetBrainsMono.zip -d /tmp/JetBrainsMono
-cp /tmp/JetBrainsMono/fonts/ttf/*.ttf ~/.local/share/fonts/
-rm -rf /tmp/JetBrainsMono
+  # JetBrains Mono Nerd Font
+  echo "Downloading latest JetBrains Mono Nerd Font..."
+  LATEST_NERD_FONT_TAG=$(curl -sL -o /dev/null -w %{url_effective} https://github.com/ryanoasis/nerd-fonts/releases/latest | rev | cut -d'/' -f1 | rev)
+  wget -O /tmp/JetBrainsMono.zip "https://github.com/ryanoasis/nerd-fonts/releases/download/${LATEST_NERD_FONT_TAG}/JetBrainsMono.zip"
+  unzip -o /tmp/JetBrainsMono.zip -d ~/.local/share/fonts
+  rm -f /tmp/JetBrainsMono.zip
 
-# Cascadia Code
-wget -O /tmp/CascadiaCode.zip https://github.com/microsoft/cascadia-code/releases/download/v2404.23/CascadiaCode-2404.23.zip
-unzip /tmp/CascadiaCode.zip -d /tmp/CascadiaCode
-cp /tmp/CascadiaCode/ttf/*.ttf ~/.local/share/fonts/
-rm -rf /tmp/CascadiaCode
+  # Cascadia Code
+  echo "Downloading latest Cascadia Code..."
+  LATEST_CASCADIA_TAG=$(curl -sL -o /dev/null -w %{url_effective} https://github.com/microsoft/cascadia-code/releases/latest | rev | cut -d'/' -f1 | rev)
+  wget -O /tmp/CascadiaCode.zip "https://github.com/microsoft/cascadia-code/releases/download/${LATEST_CASCADIA_TAG}/CascadiaCode-${LATEST_CASCADIA_TAG#v}.zip"
+  unzip -o /tmp/CascadiaCode.zip -d /tmp/CascadiaCode
+  cp /tmp/CascadiaCode/ttf/*.ttf ~/.local/share/fonts/
+  rm -rf /tmp/CascadiaCode /tmp/CascadiaCode.zip
 
-fc-cache -f -v
+  fc-cache -f -v
+else
+  echo "System is Arch, skipping font installation"
+fi
